requirements.txt
Django==1.11.5
django-cors-headers==2.1.0
django-extensions==1.9.1
django-grappelli==2.10.1
django-rest-framework==0.1.0
djangorestframework==3.6.4
factory-boy==2.9.2
Faker==0.8.7
pkg-resources==0.0.0
psycopg2==2.7.3.1
python-dateutil==2.6.1
pytz==2017.2
six==1.11.0
text-unidecode==1.1
